pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "doc23"
version = "0.1.0"
description = "Tool to extract an ordered text structure from documents"
readme = "README.md"
license = {text = "MIT"}
authors = [{ name="Alex Vargas" }]
requires-python = ">=3.10"
keywords = ["text-extraction", "document-parsing", "regex", "structure", "JSON"]
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "certifi==2025.1.31",
    "cffi==1.17.1",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "cryptography==44.0.2",
    "dnspython==2.7.0",
    "docx2txt==0.8",
    "pdf2image==1.17.0",
    "python-docx==1.1.0",
    "odfpy==1.4.1",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "Jinja2==3.1.5",
    "markdown==3.5.2",
    "python-magic==0.4.27",
    "python-multipart==0.0.20",
    "PyYAML==6.0.2",
    "sniffio==1.3.1",
    "striprtf==0.0.28",
    "typing_extensions==4.12.2",
    "pdfplumber==0.11.5",
    "pytesseract==0.3.10",
    "Pillow==11.1.0",
]

[project.urls]
Homepage = "https://github.com/alexvargashn/doc23"
Repository = "https://github.com/alexvargashn/doc23"
